[IMP] Moved page_not_found out of util.py into blueprints namespace
