+++ source/app/blueprints/pages/manage/manage_customers_routes.py
@@ -33,7 +33,7 @@ from app.models.authorization import Permissions
 from app.schema.marshables import ContactSchema
 from app.blueprints.access_controls import ac_requires
 from app.blueprints.access_controls import ac_requires_client_access
-from app.util import page_not_found
+from app.blueprints.responses import page_not_found
 from app.util import response_error
 
 manage_customers_blueprint = Blueprint(
